Truncate unnecessary data before specified offset
When we read big file, the actual readings are separated to several blocks. In
such situations, a node is requested to read at specific offset, but CRFS
doesn't truncate unnecessary data before the offset.

As a result, big files are broken in CRFS.

This commit solve this issue by truncating unnecessary data before specified
offset when CRFS fetch first chunk of required range.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>