Ensure writing to the digester completes before the digest calculation.
Sometimes, stargzifying an image fails with DIGEST_INVALID.

There is a race condition on `digester`.
The method `layer.Compressed()` calculates the digest(with `Sum()` method)
immediately after the `PipeWriter` is unblocked. But even if the `PipeWriter` is
unblocked, it isn't guaranteed that the `TeeReader` completes writing the data
to the `digester`.

This commit solves this issue by ensuring wrinting to the `digester` completes
before the digest calculation using `MultiWriter` instead of `TeeReader`.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>