Fix regression on big files from earlier optimization
An earlier change added a chunk reading cache to minimize the number of HTTP Range
requests made by the stargz ReaderAt implementation. But the use of that cache was missing
some offset adjustments when reading within a large cached chunk over multiple calls.

Fix that, and also add tests. 

Fixes #17