Reserve TOC JSON filename to avoid duplicated entries in a stargzified layer
It is possible for a layer to be stargzified twice during the distribution
lifecycle, which currently results in a broken layer. The reason is the name of
TOC JSON file("stargz.index.json") isn't reserved and duplicates in the
stargzified layer.

This commit solves this issue by reserving the TOC JSON filename when
stargzifying a layer.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>