Add an option to specify size to fetch along with stargz footer
In some cases (e.g. when high bandwidth and high latency), 2 roundtrips(for
footer + TOC json) can be a performance overhead on the mount.
This commit mitigates this issue by adding an option to specify the size to
fetch along with the stargz footer. By the option, we can hopefully get the TOC
JSON file + footer in one go, which reduces roundtrips.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>